WIP on anonymous login; split api.py and added logic.py
